revert

+++ src/uet/oop/bomberman/level/Coordinates.java
@@ -5,7 +5,7 @@ import uet.oop.bomberman.Game;
 public class Coordinates {
 	
 	public static int pixelToTile(double i) {
-		return (int)(i / Game.TILES_SIZE);
+		return (int)((int)i / Game.TILES_SIZE);
 	}
 	
 	public static int tileToPixel(int i) {
